custom videoPlayer and recorder

+++ controllers/videoController.js
@@ -36,9 +36,12 @@ export const postUpload = async (req, res) => {
 	const newVideo = await Video.create({
 		fileUrl: path,
 		title,
-		description
+		description,
+		creator: req.user.id
 	});
-	console.log(newVideo)
+	//console.log(newVideo)
+	req.user.videos.push(newVideo.id);
+	req.user.save();
 	res.redirect(routes.videoDetail(newVideo.id))
 };
 
@@ -48,7 +51,7 @@ export const videoDetail = async (req, res) => {
 		params: {id}
 	} = req;
 	try{
-		const video = await Video.findById(id);
+		const video = await (await Video.findById(id)).populate('creator');
 		console.log(video);
 		res.render("videoDetail", { pageTitle: video.title, video });
 	} catch(error) {
@@ -56,13 +59,17 @@ export const videoDetail = async (req, res) => {
 		res.redirect(routes.home);
 	}
 }
-export const getEditVideo = async (req, res) => {		// 기존 video 정보를 가져와 템플릿 랜더링
+export const getEditVideo = async (req, res) => {// 기존 video 정보를 가져와 템플릿 랜더링
 	const {
 		params: {id}		// id를 가져옴
 	} = req;
 	try{
 		const video = await Video.findById(id);	// 해당 id의 video를 가져옴
-		res.render("editVideo", { pageTitle: `Edit ${video.title}`, video })
+		if(video.creator != req.user.id){
+			throw Error();
+		} else{
+			res.render("editVideo", { pageTitle: `Edit ${video.title}`, video })
+		}
 	} catch(error){
 		res.redirect(routes.home);
 	}
@@ -86,7 +93,12 @@ export const deleteVideo = async (req, res) => {
 		params: {id}
 	} = req;
 	try{
-		await Video.findOneAndRemove({ _id: id});
+		const video = await Video.findById(id);	// 해당 id의 video를 가져옴
+		if(video.creator != req.user.id){
+			throw Error();
+		} else{
+			await Video.findOneAndRemove({ _id: id});
+		}
 	} catch(error){
 		console.log(error);
 	}
